MAJOR changes: unify-refactor monstrosity between legacy rooms and questions, work on debounce, fix textarea issue and questionID not getting re-rendered (causing unexpected overrides and strange scroll preservation behaviors)
